Moved authenticator abstraction and service to the vault folder
