+++ libs/common/src/vault/abstractions/fido2/fido2-authenticator.service.abstraction.ts
